requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
gunicorn==21.2.0
Werkzeug==2.3.7

# Database
SQLAlchemy>=2.0.30
psycopg[binary]==3.2.9
psycopg2-binary==2.9.7
alembic==1.13.1

# Caching and Performance
redis==5.0.1
Flask-Caching==2.1.0
Flask-Session==0.5.0

# HTTP and API
requests==2.31.0
psutil==5.9.6
aiohttp==3.9.1
urllib3==2.0.7

# Data processing
pandas==2.2.0
numpy==1.26.4

# Logging and monitoring
structlog==23.1.0
sentry-sdk[flask]>=2.35.0

# Environment and configuration
python-dotenv==1.0.0
pydantic>=2.8.0

# Validation and serialization
marshmallow>=3.20.1
email-validator>=2.0.0

# Web scraping and automation
playwright>=1.44.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Image processing
Pillow>=10.0.0
cloudinary>=1.36.0

# Development and testing
pytest==7.4.2
pytest-cov==4.1.0
black==25.1.0
flake8==6.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
